+++ src/stack.c
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-int stack_push(struct stack *stack, uint8_t value) {
+int stack_push(struct stack *stack, int value) {
     if (stack->pos >= STACK_SIZE) {
         return RC_ERR_STACK_OVERFLOW;
     }
@@ -13,7 +13,7 @@ int stack_push(struct stack *stack, uint8_t value) {
     return RC_SUCCESS;
 }
 
-int stack_pop(struct stack *stack, uint8_t *value) {
+int stack_pop(struct stack *stack, int *value) {
     if (stack->pos <= 0) {
         return RC_ERR_STACK_UNDERFLOW;
     }
